requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-Dance==7.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
pyngrok==6.0.0
python-dotenv==1.0.0
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
urllib3==2.0.2
URLObject==2.4.3
Werkzeug==2.3.4


